gui image display support

+++ src/main/java/thewall/engine/twilight/gui/GuiImage.java
@@ -1,17 +1,13 @@
 package thewall.engine.twilight.gui;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
 import org.joml.Vector2f;
-import org.joml.Vector3f;
 import thewall.engine.twilight.material.Material;
-import thewall.engine.twilight.spatials.Spatial;
 import thewall.engine.twilight.texture.Texture;
 
 public class GuiImage extends GUIComponent {
     private Material material;
 
-    public GuiImage(Texture texture, Vector2f transformation, int scaleX, int scaleY){
+    public GuiImage(Texture texture, Vector2f transformation, float scaleX, float scaleY){
         if(texture == null){
             throw new NullPointerException("Texture is null");
         }
